simplify(qa-smoke-file): apply remaining codex changes
